Introduce minor quality of life improvements for movement
The input movement vector is capped at a maximum magnitude of 1 to
prevent conflicts between the server and client from causing jitter.
Rotation speed has been modified to no longer depend on framerate, and
the mouse can only be used to look around when it's captured, allowing
for easier multitasking while the app is open.