Adjust near clipping plane to be much closer
The previous value of znear was too large for the default collision
sphere's radius, so it has been reduced. For future proofing, it was
reduced by two orders of magnitude. The reverse depth buffer means that
this should not cause z-fighting.

This commit also simplifies the formula for za in frustum.rs, as the old
formula is prone to catastrophic cancellation for small values of znear.
This caused visual glitches for znear=1.0e-7.