Increase precision of chunk collision checking
Relying on formulas that depend on the sinh of the radius instead of the
cosh allows sphere-casting formulas to be more precise because when the
radius is small, sinh has better precision than cosh.

This fixes noticeable precision issues in collision resolution that had
the character getting slightly jolted by the edges between adjacent
voxel faces.