Gather the server's accumulated changes into a struct
While the server is making changes to the world, it keeps a record of
these changes for one frame, which is later transmitted to exiting
clients in a Spawns struct. Rather than storing these changes ad-hoc in
several fields and local variables, this change combines them all into a
single struct to enhance readability and hopefully avoid mistakes.