Fix issues with Windows package workflow
An update to the shaderc-sys crate from 0.7 to 0.8 added a requirement
to be able to view the vk.xml file to gather information about the
installed version of Vulkan. To fix the workflow, we need to include
that file in the list of files being extracted.

In addition, the cache action was upgraded from v2 to v4 because the old
one was broken with errors like the following:
C:/VulkanSDK/1.3.231.1/: Can't create '\\\\?\\C:\\VulkanSDK\\1.3.231.1'