Run switch to 2024 edition with automatic and manual changes
Note that this doesn't have any fixes for "Tail expression temporary
scope", so there are some semantics changes. I believe this isn't an
issue that would cause a bug in Hypermine.

Manual changes:

Most of these are undoing the switch from if-let to match statements, as
the new semantics of if-let are what we wanted the whole time.

A comment "Intermediate variable to make borrowck happy" was removed
because the code also compiles without an intermediate value. This isn't
2024-edition-specific though, but it came up when checking "Tail
expression temporary scope" warnings