Improve consistency of math module's operator implementations
Use the fully qualified name instead of importing std::ops. Always
return Self::Output instead of sometimes returning Self. Implement
missing SubAssign and DivAssign.

Since we will want to implement operators for reference types as well,
the implied Self type parameter is replaced with an explicit type
parameter for all binary options to make it clear whether it's a value
type or a reference type. A similar decision was made for the associated
type `Output`.