Remove unnecessary generic features from Plane
Given that MDirection has almost everything needed to represent a plane
in usual circumstances, the Plane struct should be used for extra
functionality that worldgen would need.

This commit removes the generic type parameter, which is unneeded for
worldgen, and it makes Frustum use an MDirection directly instead of a
Plane.