Rename "length" to "depth" and make "parent" more useful
The term "depth" is arguably more intuitive when expressing how far away
from the root a graph's node is, so this should hopefully reduce the
barrier to entry for learning the Hypermine codebase.

As for "parent", previously, it was often interpreted as though every
node (except the root) has one parent, even though most logic took all
"shorter neighbors" into account, so "parent" now refers to any neighbor
node closer to the root. The old meaning of parent is renamed to
"primary parent", which is the parent with the first side in enum order.